Fixed the version reinstall error

+++ Python/InstallAndConfigurePyEnv.sh
@@ -4,7 +4,10 @@ declare PYTHON_VERSION=$2
 declare V_HOME=/home/$SSH_USERNAME
   
 sudo apt-get install -y git python-pip make build-essential libssl1.0-dev zlib1g-dev libbz2-dev libreadline-dev libsqlite3-dev
-git clone https://github.com/yyuu/pyenv.git $V_HOME/.pyenv
+
+if [ ! -f $V_HOME/.pyenv ]; then
+    git clone https://github.com/yyuu/pyenv.git $V_HOME/.pyenv
+fi
 
 sed -i '/PYENV_/d' $V_HOME/.bashrc
 sed -i '/pyenv/d' $V_HOME/.bashrc
@@ -18,11 +21,14 @@ export PATH="$PYENV_ROOT/bin:$PATH"
 eval "$(pyenv init -)"
 
 sudo chmod a+w -R $V_HOME/.pyenv
+sudo chmod a+r -R $V_HOME/.pyenv
 
-pyenv install $PYTHON_VERSION
+if [ ! -f $V_HOME/.pyenv/versions/$PYTHON_VERSION ]; then
+    pyenv install $PYTHON_VERSION
+fi
 
 pyenv global $PYTHON_VERSION
 
 unset V_HOME
-unset SSH_USERNAME
 unset PYTHON_VERSION
+unset SSH_USERNAME
